+++ R/h_map_for_count_abnormal.R
@@ -77,14 +77,14 @@ h_map_for_count_abnormal <- function(df,
   assertthat::assert_that(
     "anl" %in% names(variables),
     "split_rows" %in% names(variables),
-    is_df_with_factors(df, list(val = variables$anl)),
     is_df_with_no_na_level(
       df,
       variables = list(anl = variables$anl, split_rows = variables$split_rows), na_level = na_level
     ),
-    !any(is.na(df[variables$split_rows])),
-    is_factor_no_na(df[[variables$anl]])
+    !any(is.na(df[variables$split_rows]))
   )
+  assert_df_with_factors(df, list(val = variables$anl))
+  assert_valid_factor(df[[variables$anl]], any.missing = FALSE)
   assert_list_of_variables(variables)
   checkmate::assert_list(abnormal, types = "character", len = 2)