+++ R/h_response_subgroups.R
@@ -45,10 +45,10 @@ NULL
 #' )
 h_proportion_df <- function(rsp, arm) {
   assertthat::assert_that(
-    is.logical(rsp),
-    is_valid_factor(arm),
     is_equal_length(rsp, arm)
   )
+  checkmate::assert_logical(rsp)
+  assert_valid_factor(arm)
   non_missing_rsp <- !is.na(rsp)
   rsp <- rsp[non_missing_rsp]
   arm <- arm[non_missing_rsp]
@@ -171,10 +171,10 @@ h_proportion_subgroups_df <- function(variables,
 #' )
 h_odds_ratio_df <- function(rsp, arm, strata_data = NULL, conf_level = 0.95, method = NULL) {
   assertthat::assert_that(
-    is_valid_factor(arm),
     is_equal_length(rsp, arm),
     assertthat::are_equal(nlevels(arm), 2)
   )
+  assert_valid_factor(arm)
 
   df_rsp <- data.frame(
     rsp = rsp,
@@ -186,9 +186,9 @@ h_odds_ratio_df <- function(rsp, arm, strata_data = NULL, conf_level = 0.95, met
     strata_name <- "strata"
 
     assertthat::assert_that(
-      is_valid_factor(strata_var),
       assertthat::are_equal(length(strata_var), nrow(df_rsp))
     )
+    assert_valid_factor(strata_var)
 
     df_rsp[[strata_name]] <- strata_var
   } else {