+++ R/kaplan_meier_plot.R
@@ -200,9 +200,8 @@ g_km <- function(df,
   tte <- variables$tte
   is_event <- variables$is_event
   arm <- variables$arm
-  assertthat::assert_that(
-    is_valid_factor(df[[arm]])
-  )
+
+  assert_valid_factor(df[[arm]])
   assert_df_with_variables(df, list(tte = tte, is_event = is_event, arm = arm))
   checkmate::assert_logical(df[[is_event]], min.len = 1, any.missing = FALSE)
   checkmate::assert_numeric(df[[tte]], min.len = 1, any.missing = FALSE)