+++ R/split_cols_by_groups.R
@@ -63,10 +63,9 @@ groups_list_to_df <- function(groups_list) {
 combine_groups <- function(fct,
                            ref = NULL,
                            collapse = "/") {
-  assertthat::assert_that(
-    (is.null(ref) || is_valid_character(ref)),
-    assertthat::is.string(collapse)
-  )
+
+  checkmate::assert_string(collapse)
+  if(!is.null(ref)) checkmate::assert_character(ref, min.chars = 1, any.missing = FALSE)
   assert_character_or_factor(fct)
 
   fct <- as_factor_keep_attributes(fct)